# api-spec-fiver.yaml
openapi: 3.0.0
info:
  title: Exam Management API
  version: 1.0.0
  description: API untuk mengelola exams, questions, sections, dan passages.
servers:
  - url: http://127.0.0.1:3000/api
    description: Local development server
paths:
...
  /sections:
    get:
      summary: Mendapatkan semua sections
      responses:
        '200':
          description: Daftar sections berhasil diambil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Section'
        '500':
          description: Gagal mengambil sections
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: string
...
components:
  schemas:
...
    Section:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            type: string
        totalQuestions:
          type: number
        estimatedTime:
          type: number
...